Move EventReplay tests under tests/ and keep GPU smoke as an example.
CPU CI collects pytest tests/ without torch or a GPU; torch-free custom-init and extract_batch_context cases run there, torch/CPU cases skip, and the profile-replay script is no longer a package test_*.py.
